Moved much of VM functionality into ScriptContext.
VM is more focused on the execution of ScriptContext
ScriptContextState now allocated and holds contextes
ScriptContext now does running and "stepping"
Simplified pushContextAndRun to use the new Stack class
Opcode strings are now in OpcodeNames
Locals variables are stored and handled in each ScriptContext. Globals remain as they are.
VmStack is now a global variable STACK